Finished  implementation for 0 values TXs

+++ Sources/IotaKit/sha3/sha3.c
@@ -205,7 +205,7 @@ sha3_Update(void *priv, void const *bufIn, size_t len)
     SHA3_ASSERT(ctx->byteIndex == 0);
     
     words = len / sizeof(uint64_t);
-    tail = len - words * sizeof(uint64_t);
+    tail = (unsigned)(len - words * (size_t)sizeof(uint64_t));
     
     SHA3_TRACE("have %d full words to process", (unsigned)words);
